Backup that was found and not finshed
To finish the edge detection code, take the spline and from the polynomial function, get the value at N number of points and then send that to the curvature method. Curvature method 1 and 2 gave similar results based on what testing. I don't recall what other bugs exist in the backup. The backup is from a month ago and haven't looked at it since. apolgigies.

Notes on uses, It can only detect one edge at a time and scans from the bottom up. So you need to crop it into 1 curve at a time. The end result should be position vs curvature